% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columbus.R
\name{adjust_datetime}
\alias{adjust_datetime}
\title{Adjust datetime}
\usage{
adjust_datetime(datetime_str, type = c("to_pacific", "to_utc"))
}
\arguments{
\item{datetime_str}{Datetime string in YYYY-MM-DD HH:MM:SS format}

\item{type}{Direction for timezone adjustment: to_pacific or to_utc}
}
\description{
Adjust datetime string from UTC to pacific timezone or vice versa.
}
\examples{
# Daylight Savings Time started on 2022-11-06
adjust_datetime(c("2022-11-05 12:00:00", "2022-11-07 12:00:00"), "to_pacific")
}
