% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columbus.R
\name{get_sample_date}
\alias{get_sample_date}
\title{Get sample date}
\usage{
get_sample_date(datetime_str, threshold_time = "05:00:00")
}
\arguments{
\item{datetime_str}{Datetime string in the format YYYY-MM-DD HH:MM:SS}

\item{threshold_time}{Time string in the format HH:MM:SS}
}
\description{
Get the sample date from datetime strings based on threshold time (to accommodate sampling that extends beyond midnight).
}
\examples{
get_sample_date(c("2023-01-05 23:00:00", "2023-01-06 03:00:00", "2023-01-06 06:00:00"))
}
