% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columbus.R
\name{prep_data_file}
\alias{prep_data_file}
\title{Prep data file}
\usage{
prep_data_file(data_file)
}
\arguments{
\item{data_file}{Path to Columbus GPS data file}
}
\description{
Prep Columbus GPS data file.
}
