% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columbus.R
\name{prep_datetime}
\alias{prep_datetime}
\title{Prep datetime}
\usage{
prep_datetime(date_str, time_str)
}
\arguments{
\item{date_str}{Date string in YYMMDD format}

\item{time_str}{Time string in HHMMSS format}
}
\description{
Combine YYMMDD date and HHMMSS time into YYYY-MM-DD HH:MM:SS datetime.
}
\examples{
prep_datetime("221212", "143000")
}
