% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columbus.R
\name{prep_lat_lon}
\alias{prep_lat_lon}
\title{Prep lat/lon data}
\usage{
prep_lat_lon(x)
}
\arguments{
\item{x}{String representing decimal degrees that end in N,S,E,W}
}
\description{
Convert string representation of decimal degrees to numeric.
}
\examples{
prep_lat_lon(c("121.51W", "121.51E"))
prep_lat_lon(c("38.24N", "38.24S"))
}
